[✨🚊] Make `OPEN` and `CLOSED` commitments
**Example of open agent:**

```book
Paul Smith & Associés

PERSONA You are a company lawyer.
Your job is to provide legal advice and support to the company and its employees.
You are knowledgeable, professional, and detail-oriented.

OPEN
```

Or just:

```book
Paul Smith & Associés

PERSONA You are a company lawyer.
Your job is to provide legal advice and support to the company and its employees.
You are knowledgeable, professional, and detail-oriented.
```

**Example of closed agent:**

```book
Paul Smith & Associés

PERSONA You are a company lawyer.
Your job is to provide legal advice and support to the company and its employees.
You are knowledgeable, professional, and detail-oriented.

CLOSED
```

-   Open agents can are modified by conversation with them _(that is already implemented)_
-   Closed agents cannot be modified by conversation with them
-   By default, agents are `OPEN` unless specified as `CLOSED`
-   For the commitments create its folder in `/src/commitments` and register it in the `/src/commitments/index.ts`
-   Keep in mind the DRY _(don't repeat yourself)_ principle.
-   Add the changes into the `/changelog/_current-preversion.md`